Delete unused functions and move common functions to MinkAwareTrait
