+++ src/Drupal/DrupalExtension/Context/DrupalContext.php
@@ -3,12 +3,8 @@
 namespace Drupal\DrupalExtension\Context;
 
 use Behat\Behat\Context\TranslatableContext;
-use Behat\Behat\Hook\Scope\AfterStepScope;
 use Behat\Gherkin\Node\TableNode;
 use Behat\Mink\Element\Element;
-use Behat\Mink\Exception\Behat\Mink\Exception\UnsupportedDriverActionException;
-use Behat\Mink\Exception\DriverException;
-use Behat\Testwork\Tester\Result\TestResult;
 
 use Drupal\DrupalExtension\FeatureTrait;
 use Drupal\DrupalExtension\MinkAwareTrait;
@@ -182,72 +178,6 @@ class DrupalContext extends RawDrupalContext implements TranslatableContext
         }
         throw new \Exception(sprintf('Failed to find a row containing "%s" on the page %s', $search, $this->getSession()->getCurrentUrl()));
     }
-
-  /**
-   * See if the table rows contains specified text from a given element.
-   *
-   * @param Element $element
-   *   \Behat\Mink\Element\Element object.
-   * @param string $search
-   *   The text to search for in the table row.
-   *
-   * @return bool
-   *
-   */
-  public function hasTableRows(Element $element, string $search): bool {
-    $rows = $element->findAll('css', 'tr');
-    if (empty($rows)) {
-      return FALSE;
-    }
-    array_filter($rows, static function ($row) use ($search) {
-      if (strpos($row->getText(), $search) !== FALSE) {
-        return TRUE;
-      }
-    });
-    return FALSE;
-  }
-
-  /**
-   * Retrieve table rows containing specified text from a given element.
-   *
-   * @param Element $element
-   *   \Behat\Mink\Element\Element object.
-   * @param string $search
-   *   The text to search for in the table row.
-   *
-   * @return array of \Behat\Mink\Element\NodeElement
-   *
-   * @throws \RuntimeException
-   */
-  public function getTableRows(Element $element, $search): array {
-    $rows = $element->findAll('css', 'tr');
-    if (empty($rows)) {
-      throw new \RuntimeException(sprintf('No rows found on the page %s', $this->getSession()->getCurrentUrl()));
-    }
-    $rows = array_filter($rows, static function ($row) use ($search) {
-      if (strpos($row->getText(), $search) !== FALSE) {
-        return $row;
-      }
-    });
-    if (empty($rows)) {
-      throw new \RuntimeException(sprintf('Failed to find a row containing "%s" on the page %s', $search, $this->getSession()->getCurrentUrl()));
-    }
-    return $rows;
-  }
-
-  /**
-   * See if the element has a table.
-   *
-   * @param Element $element
-   *    \Behat\Mink\Element\Element object.
-   *
-   * @return bool
-   *   True if the element has a table.
-   */
-  private function hasTable(Element $element): bool {
-    $rows = $element->findAll('css', 'tr');
-    return !empty($rows);
-  }
 
   /**
    * Find text in a table row containing given text.
@@ -606,168 +536,5 @@ class DrupalContext extends RawDrupalContext implements TranslatableContext
         } while (true);
         fwrite(STDOUT, "\033[u");
     }
-
-  /**
-   * Wait specified milliseconds.
-   *
-   * @When I wait :msec milliseconds
-   */
-  public function waitMilliSeconds($msec): void {
-    usleep($msec * 1000);
-  }
-
-  /**
-   * Find text in a table row containing given text.
-   *
-   * @Then I should see (the text ):text1 or :text2 in the :rowText row
-   */
-  public function assertTextOrInTableRow($text1, $text2, $rowText): void {
-    $row = $this->getTableRow($this->getSession()->getPage(), $rowText);
-    if (strpos($row->getText(), $text1) === FALSE && strpos($row->getText(), $text2) === FALSE) {
-      throw new \RuntimeException(sprintf('Found a row containing "%s", but it contained neither the text "%s" nor "%s".', $rowText, $text1, $text2));
-    }
-  }
-
-  /**
-   * Check a box in a table row containing given text.
-   *
-   * @When I check the box in the :rowText row
-   */
-  public function checkBoxInTableRow($rowText): void {
-    $row = $this->getTableRow($this->getSession()->getPage(), $rowText);
-    $input = $row->find('css', 'input');
-    $input->check();
-  }
-
-  /**
-   * Fill in a form field w/ id|name|title|alt|value in the specified table row.
-   *
-   * @When I fill in :value for :field in the row :row in the :region( region)
-   */
-  public function fillFieldInTableRowInRegion($value, $field, $row, $region): void {
-    $regionObj = $this->getRegion($region);
-    $tableRowSelector = sprintf('.row-%d', intval($row));
-    $rowObj = $regionObj->find('css', $tableRowSelector);
-    $rowObj->fillField($field, $value);
-  }
-
-  /**
-   * Follows link in a table row with given value set in specified field.
-   *
-   * @When I follow/click :link in a row with :value set in :field in the :region( region)
-   */
-  public function followLinkInTableRowWithValueInRegion($link, $value, $field, $region): void {
-    $regionObj = $this->getRegion($region);
-    $rows = $regionObj->findAll('css', 'table > tbody > tr');
-    foreach ($rows as $row) {
-      $fieldObj = $row->findField($field);
-      if (strpos($fieldObj->getValue(), $value) === FALSE) {
-        continue;
-      }
-      $row->clickLink($link);
-      return;
-    }
-  }
-
-  /**
-   * Find text in a table row containing given text in specified region.
-   *
-   * @Then I should see (the text ):text in the :rowText row in the :region( region)
-   */
-  public function assertTextInTableRowInRegion($text, $rowText, $region): void {
-    $regionObj = $this->getRegion($region);
-    $row = $this->getTableRow($regionObj, $rowText);
-    if (strpos($row->getText(), $text) === FALSE) {
-      throw new \Exception(sprintf("The text '%s' was not found in the row with '%s' in the region '%s' on the page %s", $text, $rowText, $region, $this->getSession()->getCurrentUrl()));
-    }
-  }
-
-  /**
-   * Take a screenshot.
-   *
-   * @Then I take screenshot
-   */
-  public function iTakeScreenshot(): void {
-    $this->takeScreenshot(FALSE);
-  }
-
-  /**
-   * Take a screenshot on failure.
-   *
-   * @AfterStep
-   */
-  public function takeScreenshotOnFailure(AfterStepScope $scope): void {
-    // We cannot use !isPassed() as the scenario outline returns a non-PASSED code.
-    if ($scope->getTestResult()->getResultCode() === TestResult::FAILED) {
-      $this->takeScreenshot(TRUE);
-    }
-  }
-
-  /**
-   * Take screenshot or content html
-   *
-   * @var bool $on_failure
-   *   TRUE if called on failure.
-   */
-  public function takeScreenshot(bool $on_failure = FALSE): void {
-
-    $filepath = $this->getScreenshotParameter('path');
-    if (empty($filepath) || (!is_dir($filepath) && !mkdir($filepath, 0777, true) && !is_dir($filepath))) {
-       $filepath = sys_get_temp_dir();
-    }
-
-    $feature_file = $this->getFeature()->getFile();
-    [$feature_filename, $extension] = explode('.',
-      substr($feature_file, strrpos($feature_file, '/') + 1));
-    $filename = sprintf('%s_%s_%s%s', date('mdy-His'),
-      $feature_filename, $this->getStep()->getLine(),
-      $on_failure
-        ? $this->getScreenshotParameter('failure_suffix')
-        : ''
-    );
-
-    $output_filepath = "${filepath}/${filename}";
-    try {
-      $suffix = 'png';
-      $this->saveScreenshot("${filename}.${suffix}", $filepath);
-      echo "Screenshot at: ${output_filepath}.${suffix}";
-    } catch (UnsupportedDriverActionException|DriverException $e) {
-      $data = $this->getSession()->getDriver()->getContent();
-      $suffix = 'html';
-      file_put_contents("${output_filepath}.${suffix}", $data);
-      echo "Screenshot at: ${output_filepath}.${suffix}";
-    }
-  }
-
-  /**
-   * Get a screenshot parameter.
-   *
-   * @var name
-   *  Sub parameter name.
-   *
-   * @return string|null
-   *  Path.
-   */
-  private function getScreenshotParameter(string $name): ?string {
-    $parameters = $this->getDrupalParameter('screenshot');
-    return !empty($parameters[$name]) ? $parameters[$name] : NULL;
-  }
-
-  /**
-   * Log in as the existing user
-   *
-   * @Given I am logged in as user :name
-   */
-  public function iAmLoggedInAsUser($name) {
-    // Another solution using user_pass_reset_url() is independent from drush,
-    // but it works only once.
-    $base_url = $this->getMinkParameter('base_url');
-    $user_login= $this->getDriver('drush')->drush('user:login', [
-      "--name=" . $name,
-      "--no-browser",
-      "--uri=" . $base_url,
-    ]);
-    $this->getSession()->visit(trim($user_login));
-  }
 
 }
